refactor: 아이템 카테고리 타입 구현 (#9)
